backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.29.0
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
python-multipart==0.0.9
pydantic-settings==2.5.2